+++ packages/site/src/pages/visualTests/VisualTestRouter.tsx
@@ -42,6 +42,7 @@ import {
   VisualTestMenuSmallStickyPage,
 } from "./VisualTestMenuSmallPage";
 import { VisualTestModalPage } from "./VisualTestModalPage";
+import { VisualTestPagePage } from "./VisualTestPagePage";
 import { VisualTestPopoverPage } from "./VisualTestPopoverPage";
 import { VisualTestProgressBarPage } from "./VisualTestProgressBarPage";
 import { VisualTestRadioGroupPage } from "./VisualTestRadioGroupPage";
@@ -53,13 +54,13 @@ import { VisualTestSpinnerPage } from "./VisualTestSpinnerPage";
 import { VisualTestSwitchPage } from "./VisualTestSwitchPage";
 import { VisualTestTablePage } from "./VisualTestTablePage";
 import { VisualTestDataTablePage } from "./VisualTestDataTablePage";
+import { VisualTestDataTableAtomsPage } from "./VisualTestDataTableAtomsPage";
 import { VisualTestTooltipPage } from "./VisualTestTooltipPage";
 import { VisualTestTypographyPage } from "./VisualTestTypographyPage";
 import { VisualTestTabsPage } from "./VisualTestTabsPage";
 import { VisualTestToastPage } from "./VisualTestToastPage";
 import { VisualTestBannerPage } from "./VisualTestBannerPage";
 import { VisualTestThemePage } from "./VisualTestThemePage";
-import { VisualTestPagePage } from "./VisualTestPagePage";
 
 export const VisualTestRouter = () => {
   return (
@@ -68,11 +69,6 @@ export const VisualTestRouter = () => {
         path="/visual-tests/components"
         exact={true}
         component={VisualTestComponentPage}
-      />
-      <Route
-        path="/visual-tests/page"
-        exact={true}
-        component={VisualTestPagePage}
       />
       <Route
         path="/visual-tests/layout"
@@ -280,6 +276,11 @@ export const VisualTestRouter = () => {
         component={VisualTestMenuSmallStickyPage}
       />
       <Route
+        path="/visual-tests/page"
+        exact={true}
+        component={VisualTestPagePage}
+      />
+      <Route
         path="/visual-tests/popover"
         exact={true}
         component={VisualTestPopoverPage}
@@ -333,6 +334,11 @@ export const VisualTestRouter = () => {
         path="/visual-tests/data-table"
         exact={true}
         component={VisualTestDataTablePage}
+      />
+      <Route
+        path="/visual-tests/data-table-atoms"
+        exact={true}
+        component={VisualTestDataTableAtomsPage}
       />
       <Route
         path="/visual-tests/tabs"